+++ pkg/ml/ag/node.go
@@ -12,7 +12,7 @@ type Node interface {
 	// Graph returns the graph this node belongs to.
 	Graph() *Graph
 	// ID returns the ID of the node in the graph.
-	ID() int64
+	ID() int
 	// TimeStep returns the time-step associated to this node.
 	TimeStep() int
 }
